Move analytics sync client to pkg
