+++ cmd/state-offline-installer/install.go
@@ -10,7 +10,6 @@ import (
 
 	"github.com/ActiveState/cli/internal/analytics"
 	ac "github.com/ActiveState/cli/internal/analytics/constants"
-	"github.com/ActiveState/cli/internal/analytics/dimensions"
 	"github.com/ActiveState/cli/internal/assets"
 	"github.com/ActiveState/cli/internal/config"
 	"github.com/ActiveState/cli/internal/errs"
@@ -86,7 +85,7 @@ func newParams() *Params {
 }
 
 func (r *runner) Run(params *Params) (rerr error) {
-	var installerDimensions *dimensions.Values
+	var installerDimensions *analytics.Dimensions
 	defer func() {
 		if rerr == nil {
 			return
@@ -117,7 +116,7 @@ func (r *runner) Run(params *Params) (rerr error) {
 	}
 
 	namespace := project.NewNamespace(r.icfg.OrgName, r.icfg.ProjectName, "")
-	installerDimensions = &dimensions.Values{
+	installerDimensions = &analytics.Dimensions{
 		ProjectNameSpace: p.StrP(namespace.String()),
 		CommitID:         &r.icfg.CommitID,
 		Trigger:          p.StrP(target.TriggerOfflineInstaller.String()),
