+++ internal/analytics/client/blackhole/client.go
@@ -2,7 +2,6 @@ package blackhole
 
 import (
 	"github.com/ActiveState/cli/internal/analytics"
-	"github.com/ActiveState/cli/internal/analytics/dimensions"
 )
 
 type Client struct{}
@@ -13,10 +12,10 @@ func New() *Client {
 	return &Client{}
 }
 
-func (c Client) Event(category string, action string, dim ...*dimensions.Values) {
+func (c Client) Event(category string, action string, dim ...*analytics.Dimensions) {
 }
 
-func (c Client) EventWithLabel(category string, action string, label string, dim ...*dimensions.Values) {
+func (c Client) EventWithLabel(category string, action string, label string, dim ...*analytics.Dimensions) {
 }
 
 func (c Client) Wait() {
