+++ internal/analytics/dimensions/dimensions.go
@@ -1,10 +1,7 @@
 package dimensions
 
 import (
-	"encoding/json"
-	"os"
-	"strings"
-
+	"github.com/ActiveState/cli/internal/analytics"
 	"github.com/ActiveState/cli/internal/constants"
 	"github.com/ActiveState/cli/internal/errs"
 	"github.com/ActiveState/cli/internal/installation/storage"
@@ -20,31 +17,7 @@ import (
 	"github.com/ActiveState/cli/pkg/sysinfo"
 )
 
-type Values struct {
-	Version          *string
-	BranchName       *string
-	UserID           *string
-	OSName           *string
-	OSVersion        *string
-	InstallSource    *string
-	UniqID           *string
-	SessionToken     *string
-	UpdateTag        *string
-	ProjectNameSpace *string
-	OutputType       *string
-	ProjectID        *string
-	Flags            *string
-	Trigger          *string
-	Headless         *string
-	InstanceID       *string
-	CommitID         *string
-	Command          *string
-	Sequence         *int
-
-	preProcessor func(*Values) error
-}
-
-func NewDefaultDimensions(pjNamespace, sessionToken, updateTag string) *Values {
+func NewDefaultDimensions(pjNamespace, sessionToken, updateTag string) *analytics.Dimensions {
 	installSource, err := storage.InstallSource()
 	if err != nil {
 		multilog.Error("Could not detect installSource: %s", errs.Join(err, " :: ").Error())
@@ -68,7 +41,7 @@ func NewDefaultDimensions(pjNamespace, sessionToken, updateTag string) *Values {
 		osVersion = osvInfo.Version
 	}
 
-	return &Values{
+	return &analytics.Dimensions{
 		p.StrP(constants.Version),
 		p.StrP(constants.BranchName),
 		p.StrP(userIDString),
@@ -81,7 +54,7 @@ func NewDefaultDimensions(pjNamespace, sessionToken, updateTag string) *Values {
 		p.StrP(pjNamespace),
 		p.StrP(string(output.PlainFormatName)),
 		p.StrP(""),
-		p.StrP(CalculateFlags()),
+		p.StrP(analytics.CalculateFlags()),
 		p.StrP(""),
 		p.StrP(""),
 		p.StrP(instanceid.ID()),
@@ -90,132 +63,4 @@ func NewDefaultDimensions(pjNamespace, sessionToken, updateTag string) *Values {
 		p.IntP(0),
 		nil,
 	}
-}
-
-func (v *Values) Clone() *Values {
-	return &Values{
-		Version:          p.PstrP(v.Version),
-		BranchName:       p.PstrP(v.BranchName),
-		UserID:           p.PstrP(v.UserID),
-		OSName:           p.PstrP(v.OSName),
-		OSVersion:        p.PstrP(v.OSVersion),
-		InstallSource:    p.PstrP(v.InstallSource),
-		UniqID:           p.PstrP(v.UniqID),
-		SessionToken:     p.PstrP(v.SessionToken),
-		UpdateTag:        p.PstrP(v.UpdateTag),
-		ProjectNameSpace: p.PstrP(v.ProjectNameSpace),
-		OutputType:       p.PstrP(v.OutputType),
-		ProjectID:        p.PstrP(v.ProjectID),
-		Flags:            p.PstrP(v.Flags),
-		Trigger:          p.PstrP(v.Trigger),
-		Headless:         p.PstrP(v.Headless),
-		InstanceID:       p.PstrP(v.InstanceID),
-		CommitID:         p.PstrP(v.CommitID),
-		Command:          p.PstrP(v.Command),
-		Sequence:         p.PintP(v.Sequence),
-		preProcessor:     v.preProcessor,
-	}
-}
-
-func (m *Values) Merge(mergeWith ...*Values) {
-	// This is awkward and long, but using mergo was not an option here because it cannot differentiate between
-	// falsy values and nil pointers
-	for _, dim := range mergeWith {
-		if dim.Version != nil {
-			m.Version = dim.Version
-		}
-		if dim.BranchName != nil {
-			m.BranchName = dim.BranchName
-		}
-		if dim.UserID != nil {
-			m.UserID = dim.UserID
-		}
-		if dim.OSName != nil {
-			m.OSName = dim.OSName
-		}
-		if dim.OSVersion != nil {
-			m.OSVersion = dim.OSVersion
-		}
-		if dim.InstallSource != nil {
-			m.InstallSource = dim.InstallSource
-		}
-		if dim.UniqID != nil {
-			m.UniqID = dim.UniqID
-		}
-		if dim.SessionToken != nil {
-			m.SessionToken = dim.SessionToken
-		}
-		if dim.UpdateTag != nil {
-			m.UpdateTag = dim.UpdateTag
-		}
-		if dim.ProjectNameSpace != nil {
-			m.ProjectNameSpace = dim.ProjectNameSpace
-		}
-		if dim.OutputType != nil {
-			m.OutputType = dim.OutputType
-		}
-		if dim.ProjectID != nil {
-			m.ProjectID = dim.ProjectID
-		}
-		if dim.Flags != nil {
-			m.Flags = dim.Flags
-		}
-		if dim.Trigger != nil {
-			m.Trigger = dim.Trigger
-		}
-		if dim.Headless != nil {
-			m.Headless = dim.Headless
-		}
-		if dim.InstanceID != nil {
-			m.InstanceID = dim.InstanceID
-		}
-		if dim.CommitID != nil {
-			m.CommitID = dim.CommitID
-		}
-		if dim.Command != nil {
-			m.Command = dim.Command
-		}
-		if dim.Sequence != nil {
-			m.Sequence = dim.Sequence
-		}
-		if dim.preProcessor != nil {
-			m.preProcessor = dim.preProcessor
-		}
-	}
-}
-
-func (v *Values) RegisterPreProcessor(f func(*Values) error) {
-	v.preProcessor = f
-}
-
-func (v *Values) PreProcess() error {
-	if v.preProcessor != nil {
-		if err := v.preProcessor(v); err != nil {
-			return errs.Wrap(err, "PreProcessor failed: %s", errs.JoinMessage(err))
-		}
-	}
-
-	if p.PStr(v.UniqID) == "" {
-		return errs.New("device id is unset when creating analytics event")
-	}
-
-	return nil
-}
-
-func (v *Values) Marshal() (string, error) {
-	dimMarshalled, err := json.Marshal(v)
-	if err != nil {
-		return "", errs.Wrap(err, "Could not marshal dimensions")
-	}
-	return string(dimMarshalled), nil
-}
-
-func CalculateFlags() string {
-	flags := []string{}
-	for _, arg := range os.Args {
-		if strings.HasPrefix(arg, "-") {
-			flags = append(flags, arg)
-		}
-	}
-	return strings.Join(flags, " ")
 }
