+++ pkg/platform/analytics/sync/client_test.go
@@ -4,47 +4,47 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/ActiveState/cli/internal/analytics/dimensions"
+	"github.com/ActiveState/cli/internal/analytics"
 	"github.com/ActiveState/cli/internal/rtutils/p"
 )
 
 func Test_mergeDimensions(t *testing.T) {
 	type args struct {
-		target *dimensions.Values
-		dims   []*dimensions.Values
+		target *analytics.Dimensions
+		dims   []*analytics.Dimensions
 	}
 	tests := []struct {
 		name string
 		args args
-		want *dimensions.Values
+		want *analytics.Dimensions
 	}{
 		{
 			"Sequence favours source",
 			args{
-				&dimensions.Values{
+				&analytics.Dimensions{
 					Sequence: p.IntP(10),
 				},
-				[]*dimensions.Values{
+				[]*analytics.Dimensions{
 					{
 						Sequence: p.IntP(100),
 					},
 				},
 			},
-			&dimensions.Values{Sequence: p.IntP(100)},
+			&analytics.Dimensions{Sequence: p.IntP(100)},
 		},
 		{
 			"Sequence favours source and accepts 0 value",
 			args{
-				&dimensions.Values{
+				&analytics.Dimensions{
 					Sequence: p.IntP(10),
 				},
-				[]*dimensions.Values{
+				[]*analytics.Dimensions{
 					{
 						Sequence: p.IntP(0),
 					},
 				},
 			},
-			&dimensions.Values{Sequence: p.IntP(0)},
+			&analytics.Dimensions{Sequence: p.IntP(0)},
 		},
 	}
 	for _, tt := range tests {
